gold/AI_Employee_Vault/requirements.txt
schedule==1.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
playwright==1.40.0
watchdog==4.0.1
psutil==5.9.8